# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MaintSys API",
        "description": "API de gerenciamento de manutenção industrial - MaintSys",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/machines": {
            "get": {
                "tags": [
                    "Máquinas"
                ],
                "summary": "Lista todas as máquinas",
                "operationId": "5998a2cdaadf6f32ab905c78c5c60f12",
                "responses": {
                    "200": {
                        "description": "Lista de máquinas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "serial_number": {
                                                "type": "string",
                                                "example": "SN-001"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Torno CNC"
                                            },
                                            "model": {
                                                "type": "string",
                                                "example": "CNC-500"
                                            },
                                            "location": {
                                                "type": "string",
                                                "example": "Galpão A"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active",
                                                "enum": [
                                                    "active",
                                                    "inactive",
                                                    "maintenance"
                                                ]
                                            },
                                            "installed_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2023-01-15"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Máquinas"
                ],
                "summary": "Cadastra uma nova máquina",
                "operationId": "9c7c1a200dac4400582d7552961c1f20",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "serial_number",
                                    "name"
                                ],
                                "properties": {
                                    "serial_number": {
                                        "type": "string",
                                        "example": "SN-002"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Fresadora"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "FR-300"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Galpão B"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "maintenance"
                                        ]
                                    },
                                    "installed_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Máquina criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "serial_number": {
                                            "type": "string",
                                            "example": "SN-002"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Fresadora"
                                        },
                                        "model": {
                                            "type": "string",
                                            "example": "FR-300"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Galpão B"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "installed_at": {
                                            "type": "string",
                                            "example": "2024-03-10"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/machines/{id}": {
            "get": {
                "tags": [
                    "Máquinas"
                ],
                "summary": "Retorna uma máquina específica",
                "operationId": "7b2a92bd0a873fb1adb329e771d9b59b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da máquina",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da máquina",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "serial_number": {
                                            "type": "string",
                                            "example": "SN-001"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Torno CNC"
                                        },
                                        "model": {
                                            "type": "string",
                                            "example": "CNC-500"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Galpão A"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "active"
                                        },
                                        "installed_at": {
                                            "type": "string",
                                            "example": "2023-01-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Máquina não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Máquinas"
                ],
                "summary": "Atualiza uma máquina",
                "operationId": "6495cb40b92d773ef7758df26fedbc7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da máquina",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "serial_number": {
                                        "type": "string",
                                        "example": "SN-001-A"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Torno CNC Atualizado"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "CNC-600"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Galpão C"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "maintenance",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "maintenance"
                                        ]
                                    },
                                    "installed_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Máquina atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "serial_number": {
                                            "type": "string",
                                            "example": "SN-001-A"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Torno CNC Atualizado"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "maintenance"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Máquina não encontrada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Máquinas"
                ],
                "summary": "Remove uma máquina",
                "operationId": "9cbc00780e4380f6622bb0589dce0f17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da máquina",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Máquina removida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Máquina removida"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Máquina não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}